rename salespeople dialog directives
